requirements.txt
asgiref==3.5.2
Django==3.2.16
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
typing-extensions==4.4.0
